Reprice to $0.50, and let customers hear their own call
Paystack refuses very small amounts outright - a 1 naira charge comes back
as "No active channel to process transaction" - so the price is 750 naira,
about $0.50, which is verified to open a real checkout.

Call recordings are now the customer's to listen to. Twilio holds the
original behind account credentials the browser must never see, and drops
recordings when an account lapses, so each one is copied into a private
bucket keyed by owner and negotiation as it arrives. Playback goes through
an ownership check before a link is minted at all, and the link is signed
and expires - a recorded phone call about somebody's account should stop
being reachable, not stay permanently addressable.

Archiving failure never blocks the verification that follows, because that
is the part a saving depends on.

Verified against the real bucket: the upload stores, the signed link plays
back the exact bytes, and an unsigned fetch of the same object is refused.

+++ backend/app/routers/billing_plan.py
@@ -36,7 +36,7 @@ class PlanState(BaseModel):
     used: int
     remaining: int | None
     month: str
-    price_usd: int
+    price_usd: float
     expires_at: str | None
     can_upgrade: bool
     # So a renewal is never a surprise, and cancelling is possible from the app.
